i2c: msm-geni: Prevent transfer when bus recovery is disabled
Previously, if the IO lines were in a bad state and bus recovery was
disabled, the driver returned success and proceeded with the transfer,
leading to timeouts.

This change adds a check for bus recovery enablement. If recovery is
disabled, the transfer is aborted and an error is returned instead of
attempting recovery. This prevents unnecessary timeouts and improves
robustness when the bus cannot be recovered.

Change-Id: Ie34ac08e63c8b880d27cf443021ba6dde01977a9
Signed-off-by: Panicker Harish <quic_pharish@quicinc.com>
Signed-off-by: Prasanna S <quic_prass@quicinc.com>
Signed-off-by: Anup Kulkarni <anup.kulkarni@oss.qualcomm.com>